feat: override "alg" option
Fixes #3

As mentioned by #3 we were not following the specification
properly. The "alg" claim is not mandatory and so I've added an option
for this claim to be passed.

Also, added integration tests to be sure we can parse those keys and
added some extra docs with the things we need in order to parse keys.

Extra tests for some tricky cases like synchronous first fetch along
with polling and so on.

Updated deps too :)